requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
PyJWT==2.8.0
pyodbc==5.1.0
python-dotenv==1.0.1
requests==2.32.3